+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 **Added:**
 * Experience boost will now boost the amount of vanilla EXP obtained when HP is 50% or below.
 * Going into a Drive Form will refill the MP indicator as well.
+* Moogle exp viewer section (WIP, needs scrollbar on materials)
 
 **Fixed:**
 * Airstep not working.
 * Fortune will now work on Prize Blox and Rare Prize Blox.
 * Single shotlock selector and airstep icon not rendering properly.
 * Synthesis bags will no longer be clickable from a synthesis bag inventory screen. 
+* Magic Command Menu being too dark.